requirements.txt
asgiref==3.7.2
boto3==1.28.60
botocore==1.31.60
certifi==2023.7.22
charset-normalizer==3.3.0
crispy-bootstrap5==0.7
Django==4.2.4
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-storages==1.14.1
docopt==0.6.2
idna==3.4
jmespath==1.0.1
Pillow==10.0.0
pipreqs==0.4.13
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.17
yarg==0.1.9
